Update README.md
This project uses a machine learning model to predict the likelihood of lung cancer based on patient symptoms and habits. By analyzing features such as smoking, age, fatigue, and chest pain, the model helps in early detection of lung cancer. It uses the Random Forest algorithm for accurate and efficient predictions, aiming to support research and awareness in health analytics.